+++ CarDealership/Classes(MiddleLayer)/Nissan.cs
@@ -10,7 +10,7 @@ namespace CarDealership
     {
         public string Transmission { get; set; }
 
-        public Nissan(string make, string model, string color, int year, decimal price, string transmisson, DateTime dateAdded)
+        public Nissan(string make, string model, string color, int year, int price, string transmisson, DateTime dateAdded)
             : base(make, model, color, year, price, dateAdded)
         {
             Transmission = transmisson;
